% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusting.h.R
\name{adjusting}
\alias{adjusting}
\title{Adjusting}
\usage{
adjusting(
  data,
  dep = NULL,
  factors,
  covs = NULL,
  model_type = "lm",
  covsTransformations = list("lin", "ln", "log100", "rec"),
  select = TRUE,
  method = "step",
  direction = "high",
  forced = NULL,
  included = NULL,
  es_zscore = TRUE,
  es_rank = FALSE,
  perc = TRUE,
  perc_type = "eby5",
  score_adjust = FALSE,
  score_raw = FALSE,
  score_preds = FALSE,
  score_sort = "none",
  plots = list(),
  plot_es = FALSE,
  .caller = "scoring",
  .interface = "jamovi"
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{Update this when R package is to be dealt with}

\item{factors}{Update this when R package is to be dealt with}

\item{covs}{Update this when R package is to be dealt with}

\item{model_type}{.}

\item{covsTransformations}{Update this when R package is to be dealt with}

\item{select}{.}

\item{method}{Update this when R package is to be dealt with}

\item{direction}{.}

\item{forced}{a named vector of the form \code{c(var1="type",
var2="type2")}}

\item{included}{.}

\item{es_zscore}{.}

\item{es_rank}{.}

\item{perc}{.}

\item{perc_type}{.}

\item{score_adjust}{.}

\item{score_raw}{.}

\item{score_preds}{.}

\item{score_sort}{.}

\item{plots}{.}

\item{plot_es}{.}

\item{.caller}{.}

\item{.interface}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$info} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$varstab} \tab \tab \tab \tab \tab a table \cr
  \code{results$final} \tab \tab \tab \tab \tab a table \cr
  \code{results$univariate} \tab \tab \tab \tab \tab a table \cr
  \code{results$multiple} \tab \tab \tab \tab \tab a table \cr
  \code{results$plots$adj} \tab \tab \tab \tab \tab an image \cr
  \code{results$plots$adj_obs} \tab \tab \tab \tab \tab an image \cr
  \code{results$plots$adj_pred} \tab \tab \tab \tab \tab an array \cr
  \code{results$plots$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$scores$formula} \tab \tab \tab \tab \tab a html \cr
  \code{results$scores$es} \tab \tab \tab \tab \tab a table \cr
  \code{results$scores$percentiles} \tab \tab \tab \tab \tab a table \cr
  \code{results$scores$raws} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$varstab$asDF}

\code{as.data.frame(results$varstab)}
}
\description{
Something here
}
\examples{
\donttest{
some code here
}
}
